LayoutTests/fast/dom/MicroData/item-not-attached-to-domtree.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/microdata-common.js"></script>
</head>
<body>
<script>
description("Test to ensure that HTMLPropertiesCollection contains all item properties when item is not attached to the DOM tree.");
var parent = createElement('div', {}, '<div itemprop="foo" id="id1"></div><div itemscope itemref="id1 id2"><div itemprop="bar"></div></div><div itemprop="foo" id="id2"><div itemprop="baz"></div></div>');
var item = parent.childNodes[1];

function executeTest() {
    shouldBe("item.properties.length", "4");
    shouldBe("item.properties.item(0)", "parent.firstChild");
    shouldBe("item.properties[0]", "parent.firstChild");
    shouldBe("item.properties.item(1)", "item.firstChild");
    shouldBe("item.properties[1]", "item.firstChild");
    shouldBe("item.properties.item(2)", "parent.childNodes[2]");
    shouldBe("item.properties[2]", "parent.childNodes[2]");
    shouldBe("item.properties.item(3)", "parent.childNodes[2].firstChild");
    shouldBe("item.properties[3]", "parent.childNodes[2].firstChild");
    shouldBe("item.properties.names.length", "3");
    shouldBe("item.properties.names.item(0)", "'foo'");
    shouldBe("item.properties.names[0]", "'foo'");
    shouldBe("item.properties.names.item(1)", "'bar'");
    shouldBe("item.properties.names[1]", "'bar'");
    shouldBe("item.properties.names.item(2)", "'baz'");
    shouldBe("item.properties.names[2]", "'baz'");
    shouldBe("item.properties.namedItem('foo').length", "2");
    shouldBe("item.properties.namedItem('foo').item(0)", "parent.firstChild");
    shouldBe("item.properties.namedItem('foo')[0]", "parent.firstChild");
    shouldBe("item.properties.namedItem('foo').item(1)", "parent.childNodes[2]");
    shouldBe("item.properties.namedItem('foo')[1]", "parent.childNodes[2]");
    shouldBe("item.properties.namedItem('bar').length", "1");
    shouldBe("item.properties.namedItem('bar').item(0)", "item.firstChild");
    shouldBe("item.properties.namedItem('bar')[0]", "item.firstChild");
    shouldBe("item.properties.namedItem('baz').length", "1");
    shouldBe("item.properties.namedItem('baz').item(0)", "parent.childNodes[2].firstChild");
    shouldBe("item.properties.namedItem('baz')[0]", "parent.childNodes[2].firstChild");
}

debug("Item not attached to the dom tree");
executeTest();

document.body.appendChild(parent);
debug("");
debug("Item added to the dom tree");
executeTest();

document.body.removeChild(parent);
debug("");
debug("Item detached from the dom tree");
executeTest();
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
